AddeD: Support for add extra kid
